avoid size_t/int64_t conversion in minimum of walk

+++ src/stats.hpp
@@ -137,7 +137,7 @@ struct Stats {
     int64_t count = 0;
     int64_t broken = 0;
     int64_t flips = 0;
-    int64_t minimum = 0;
+    size_t minimum = 0;
     int64_t improved = 0;
   } walk;
